+++ crates/qpy/src/value.rs
@@ -123,7 +123,7 @@ pub(crate) fn unpack_biguint(big_int_pack: BigIntPack) -> BigUint {
 #[derive(Debug)]
 pub struct QPYWriteData<'a> {
     pub circuit_data: &'a CircuitData,
-    pub version: u32,
+    pub version: u8,
     pub standalone_var_indices: HashMap<u128, u16>, // mapping from the variable's UUID to its index in the standalone variables list
     pub annotation_handler: AnnotationHandler<'a>,
 }
@@ -132,7 +132,7 @@ pub struct QPYWriteData<'a> {
 #[derive(Debug)]
 pub struct QPYReadData<'a> {
     pub circuit_data: &'a mut CircuitData,
-    pub version: u32,
+    pub version: u8,
     pub use_symengine: bool,
     pub standalone_vars: HashMap<u16, qiskit_circuit::Var>,
     pub standalone_stretches: HashMap<u16, qiskit_circuit::Stretch>,
@@ -540,7 +540,7 @@ pub(crate) fn load_value(
         }
         ValueType::Circuit => {
             let (packed_circuit, _) =
-                deserialize_with_args::<formats::QPYCircuit, (u32,)>(bytes, (qpy_data.version,))?;
+                deserialize_with_args::<formats::QPYCircuit, (u8,)>(bytes, (qpy_data.version,))?;
             Python::attach(|py| {
                 let circuit = unpack_circuit(
                     py,
@@ -812,7 +812,7 @@ pub(crate) fn deserialize_expression(
 pub(crate) fn pack_standalone_var(
     var: &Var,
     usage: ExpressionVarDeclaration,
-    version: u32,
+    version: u8,
     uuid_output: &mut u128,
 ) -> Result<formats::ExpressionVarDeclarationPack, QpyError> {
     match var {
@@ -848,7 +848,7 @@ pub(crate) fn pack_stretch(
 
 // we convert the type to the serializable struct; this amounts to simple copy unless
 // there's a field not supported in the expected version
-fn pack_expression_type(exp_type: &Type, version: u32) -> Result<ExpressionType, QpyError> {
+fn pack_expression_type(exp_type: &Type, version: u8) -> Result<ExpressionType, QpyError> {
     match exp_type {
         Type::Bool => Ok(ExpressionType::Bool),
         Type::Duration => {
